Added code to skip already solved boards
I had commented out code that kept track of the board in an unordered
set and, if found, skipped the solving of the board. However I realized
this doesn't work because the solution could arrive at a board with a
higher score. This submission I witched to an unordered map and store
the score as well,  I ran this for a few minutes with some test outputs
and it eliminates a large chunk of the time. Although I still have yet
to get it to the point to where I can run it till its finished (I've had
it running in the background for a few hours and although it's processed
billions of solutions, it's still not done). Again, you get a good
answer in seconds and sometimes a better one in minutes. Occasionally
you'll get an improvement an hour in or so.